+++ src/bambu_bridge/log_redaction.py
@@ -10,7 +10,7 @@ from urllib.parse import quote, quote_plus
 
 _QUERY = re.compile(r'''(?i)([?&](?:token|api_key|access_token|access_code|key)=)[^&\s\#'"<>]*''')
 _BEARER = re.compile(r'''(?i)(\bBearer\s+)[^\s'"<>]+''')
-_SECRET_FIELDS = {"token", "api_key", "bridge_api_key", "bridge_viz_token", "access_code"}
+_SECRET_FIELDS = {"token", "api_key", "bridge_api_key", "bridge_viz_token", "access_code", "secret"}
 _secrets: frozenset[str] = frozenset()
 _original_factory = logging.getLogRecordFactory()